+++ services/src/main/java/io/scalecube/services/ServiceRegistryImpl.java
@@ -3,6 +3,7 @@ package io.scalecube.services;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import io.scalecube.cluster.Member;
+import io.scalecube.concurrency.ThreadFactory;
 import io.scalecube.services.ServicesConfig.Builder.ServiceConfig;
 import io.scalecube.services.metrics.Metrics;
 import io.scalecube.transport.Address;
