+++ utils/src/main/java/io/scalecube/concurrency/Futures.java
@@ -1,0 +1,34 @@
+package io.scalecube.concurrency;
+
+
+import java.time.Duration;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+public class Futures {
+
+  /**
+   * used to complete the request future with timeout exception in case no response comes from service.
+   */
+  private static final ScheduledExecutorService delayer =
+      ThreadFactory.singleScheduledExecutorService(ThreadFactory.SC_SERVICES_TIMEOUT);
+  
+  public static <T> CompletableFuture<T> withTimeout(final CompletableFuture<T> resultFuture,
+      Duration timeout) {
+
+    if (!resultFuture.isDone()) {
+      // schedule to terminate the target goal in future in case it was not done yet
+      delayer.schedule(() -> {
+        // by this time the target goal should have finished.
+        if (!resultFuture.isDone()) {
+          // target goal not finished in time so cancel it with timeout.
+          resultFuture.completeExceptionally(new TimeoutException("expecting response reached timeout!"));
+        }
+      }, timeout.toMillis(), TimeUnit.MILLISECONDS);
+    }
+    return resultFuture;
+  }
+  
+}
